Struct crossfont::ft::fc::pattern::PatternRef

pub struct PatternRef(/* private fields */);

A borrowed reference to a Pattern.

Implementations

impl PatternRef

pub fn antialias(&self) -> BooleanPropertyIter<'_>

pub fn hinting(&self) -> BooleanPropertyIter<'_>

pub fn verticallayout(&self) -> BooleanPropertyIter<'_>

pub fn autohint(&self) -> BooleanPropertyIter<'_>

pub fn globaladvance(&self) -> BooleanPropertyIter<'_>

pub fn scalable(&self) -> BooleanPropertyIter<'_>

pub fn symbol(&self) -> BooleanPropertyIter<'_>

pub fn color(&self) -> BooleanPropertyIter<'_>

pub fn minspace(&self) -> BooleanPropertyIter<'_>

pub fn embolden(&self) -> BooleanPropertyIter<'_>

pub fn embeddedbitmap(&self) -> BooleanPropertyIter<'_>

pub fn decorative(&self) -> BooleanPropertyIter<'_>

pub fn size(&self) -> DoublePropertyIter<'_>

pub fn aspect(&self) -> DoublePropertyIter<'_>

pub fn pixelsize(&self) -> DoublePropertyIter<'_>

pub fn pixelsizefixupfactor(&self) -> DoublePropertyIter<'_>

pub fn scale(&self) -> DoublePropertyIter<'_>

pub fn dpi(&self) -> DoublePropertyIter<'_>

pub fn add_family(&mut self, value: &str) -> bool

pub fn family(&self) -> StringPropertyIter<'_>

pub fn add_familylang(&mut self, value: &str) -> bool

pub fn familylang(&self) -> StringPropertyIter<'_>

pub fn add_style(&mut self, value: &str) -> bool

pub fn style(&self) -> StringPropertyIter<'_>

pub fn add_stylelang(&mut self, value: &str) -> bool

pub fn stylelang(&self) -> StringPropertyIter<'_>

pub fn add_fullname(&mut self, value: &str) -> bool

pub fn fullname(&self) -> StringPropertyIter<'_>

pub fn add_fullnamelang(&mut self, value: &str) -> bool

pub fn fullnamelang(&self) -> StringPropertyIter<'_>

pub fn add_foundry(&mut self, value: &str) -> bool

pub fn foundry(&self) -> StringPropertyIter<'_>

pub fn add_capability(&mut self, value: &str) -> bool

pub fn capability(&self) -> StringPropertyIter<'_>

pub fn add_fontformat(&mut self, value: &str) -> bool

pub fn fontformat(&self) -> StringPropertyIter<'_>

pub fn add_fontfeatures(&mut self, value: &str) -> bool

pub fn fontfeatures(&self) -> StringPropertyIter<'_>

pub fn add_namelang(&mut self, value: &str) -> bool

pub fn namelang(&self) -> StringPropertyIter<'_>

pub fn add_postscriptname(&mut self, value: &str) -> bool

pub fn postscriptname(&self) -> StringPropertyIter<'_>

pub fn index(&self) -> IntPropertyIter<'_>

pub fn print(&self)

Prints the pattern to stdout.

FontConfig doesn’t expose a way to iterate over all members of a pattern; instead, we just defer to FcPatternPrint. Otherwise, this could have been a fmt::Debug impl.

pub fn hintstyle(&self) -> HintStylePropertyIter<'_>

pub fn lcdfilter(&self) -> LcdFilterPropertyIter<'_>

pub fn set_slant(&mut self, slant: Slant) -> bool

pub fn add_pixelsize(&mut self, size: f64) -> bool

pub fn set_weight(&mut self, weight: Weight) -> bool

pub fn set_width(&mut self, width: Width) -> bool

pub fn get_width(&self) -> Option<Width>

pub fn rgba(&self) -> RgbaPropertyIter<'_>

pub fn set_rgba(&self, rgba: &Rgba) -> bool

pub fn render_prepare( &self, config: &ConfigRef, request: &PatternRef ) -> Pattern

pub fn hash(&self) -> PatternHash

pub fn add_charset(&self, charset: &CharSetRef) -> bool

Add charset to the pattern.

The referenced charset is copied by Fontconfig internally using FcValueSave so that no references to application provided memory are retained. That is, the CharSet can be safely dropped immediately after being added to the pattern.

pub fn get_charset(&self) -> Option<&CharSetRef>

Get charset from the pattern.

pub fn get_matrix(&self) -> Option<FcMatrix>

Get matrix from the pattern.

pub fn file(&self, index: usize) -> Option<PathBuf>

pub fn ft_face_location(&self, index: usize) -> Option<FtFaceLocation>

pub fn config_substitute(&mut self, config: &ConfigRef, kind: MatchKind)

pub fn default_substitute(&mut self)